Big update to .... Everything
Ok I finally have most of the bugs worked out. Now I'm trying to prevent the GA from cheating in path length due to exponential fitness gains in network connectivity.  Its the only way I could think of to couple Connectivity with path score and have an increase in connectivity have meaning. Anyways I added a visualizer that can be switched on and of by clicking. So you can watch the path or check on things which slows it down a lot or turn it off and get the full speed evaluation.  So far things look promising. I have some updates I'm working on to MEGA which will enhance this as well but I'm not going into detail yet just know its on its way.