Update to everything
There were a few systemic errors that were getting in the way of everything working. However I managed to get it to build the network up to full connectivity a couple times. I think theres more room to tweak things parameters but I think this is at least solid proof that we can use MEGA in this way to wire the Network. Phase one. Moving on to try to build phase two Which might require some core modifications of the SP_NN since methods have changed somewhat. The concept still works just a slightly different approach to getting it all wired together.