Created file IDK
IDK contains the original proof of concept for the SPNN its much simpler but uses the same premise of relocating objects in the environment to improve fitness. Need further study to see what Im missing because their consistent positive results show that the siring of SPNN should be possible despite it not quite working yet.