Update SP_NN.py
Added in compute_structureal_density and compute_network _health as a means of combining the structural connectivity score which is a bit frail with a scaled measure of how densely connected the network as a whole is. This should give a Higher quality fitness signal to the GA to work with and hopefully solves our wiring problem.