Big progress update
Ok so the path was the major issue. In the proof of concept the path was a discrete set of steps and they would evolve to increase in complexity. The issue was that the steps were discrete so the agent couldnt step into continuous space between the grid lines. Originally I set up a binary encoding system for the path to allow it to have a continuous coordinate system. This helped a bit but the results were still in a random range and couldn't make progress. Then it clicked that binary encodings went increasing in complexity. So I went back to the  grid based approach but instead introduced three new genes Scale up, Scale down, Toggle pick up. This way the Path can evolve different scales while still being modular so it can scale the step size up and down and travel through the network without picking up neurons.  Its not right yet but its a huge step forward as the network can go from barely connected to around 70% fairly easily.