New update
There was an issue with the Path degrading down to the lowest scale after clearing the area local to the origin since the path is now anchored to the origin. Resulting in the GA playing it safe better to get the max path score and do nothing else because it harms fitness. I introduced  a reward per step and a delta reward that rewards positive changes in network health and penalizes negative changes in health.. I also introduced a counter and if health doesn't change by a set percentage then the path reward begins to be worth less each evaluation until there is an acceptable improvement.  This seems to make it work by eliminating the safety of doing nothing.